Added basic card functionality
So far I have made a draggable card that inherits information from a repeatable dataset.

The card drags according to the mouse movement but it does not lock on to the cursor and the cursor disappears.

The dataset script is used to create "card" objects which each have their own data, this data can then be applied to a premade prefab.
Once the data is inherited it is then displayed on the card.